Fix: 푼 퀴즈 불러오기
